Move bytes_to_hex to encoding utils package.
