+++ mycli/encodingutils.py
@@ -1,7 +1,9 @@
+import binascii
 import sys
 
 PY2 = sys.version_info[0] == 2
 PY3 = sys.version_info[0] == 3
+
 
 def unicode2utf8(arg):
     """
@@ -13,6 +15,7 @@ def unicode2utf8(arg):
         return arg.encode('utf-8')
     return arg
 
+
 def utf8tounicode(arg):
     """
     Only in Python 2. Psycopg2 returns the error message as utf-8.
@@ -22,3 +25,21 @@ def utf8tounicode(arg):
     if PY2 and isinstance(arg, str):
         return arg.decode('utf-8')
     return arg
+
+
+def bytes_to_hex(b):
+    """Convert bytes that cannot be decoded to utf8 to hexlified string.
+
+    >>> print(bytes_to_hex(b"\\xff"))
+    0xff
+    >>> print(bytes_to_hex('abc'))
+    abc
+    >>> print(bytes_to_hex('✌'))
+    ✌
+    """
+    if isinstance(b, bytes):
+        try:
+            b.decode('utf8')
+        except:
+            b = '0x' + binascii.hexlify(b).decode('ascii')
+    return b
